feat: add storage interface
