+++ internal/storage/pebble/storage_test.go
@@ -1,65 +1,66 @@
-package storage
+package pebble
 
 import (
 	"os"
 	"strconv"
 	"testing"
 
+	"github.com/Taraxa-project/taraxa-indexer/internal/storage"
 	"github.com/Taraxa-project/taraxa-indexer/models"
 	"github.com/nleeper/goment"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestGetter(t *testing.T) {
-	storage := NewStorage("")
-	defer storage.Close()
+	st := NewStorage("")
+	defer st.Close()
 
-	addr := MakeEmptyAddressStats("test")
-	if err := storage.addToDBTest(addr, addr.Address, 0); err != nil {
+	addr := storage.MakeEmptyAddressStats("test")
+	if err := st.addToDBTest(addr, addr.Address, 0); err != nil {
 		t.Error(err)
 	}
-	addr1 := MakeEmptyAddressStats("test1")
-	if err := storage.addToDBTest(addr1, addr1.Address, 0); err != nil {
+	addr1 := storage.MakeEmptyAddressStats("test1")
+	if err := st.addToDBTest(addr1, addr1.Address, 0); err != nil {
 		t.Error(err)
 	}
-	ret := storage.GetAddressStats("test")
-	if !ret.isEqual(addr) {
+	ret := st.GetAddressStats("test")
+	if !ret.IsEqual(addr) {
 		t.Error("Broken DB")
 	}
 }
 
 func TestGetObjects(t *testing.T) {
-	stor := NewStorage("")
-	defer stor.Close()
+	st := NewStorage("")
+	defer st.Close()
 
 	sender := "user"
 	count := uint64(100)
 
-	stats := MakeEmptyAddressStats(sender)
+	stats := storage.MakeEmptyAddressStats(sender)
 	stats.DagsCount = count
-	if err := stor.addToDBTest(stats, stats.Address, 0); err != nil {
+	if err := st.addToDBTest(stats, stats.Address, 0); err != nil {
 		t.Error(err)
 	}
 
 	for i := uint64(1); i <= count; i++ {
 		block := models.Dag{Timestamp: i, Hash: "test" + strconv.FormatUint(i, 10), Level: 0, Sender: sender, TransactionCount: 0}
-		if err := stor.addToDBTest(&block, sender, block.Timestamp); err != nil {
+		if err := st.addToDBTest(&block, sender, block.Timestamp); err != nil {
 			t.Error(err)
 		}
 	}
-	ret, pagination := GetObjectsPage[models.Dag](stor, sender, 0, uint64(count))
+	ret, pagination := storage.GetObjectsPage[models.Dag](st, sender, 0, uint64(count))
 	assert.Equal(t, uint64(len(ret)), count)
 	assert.False(t, pagination.HasNext)
 	assert.Equal(t, pagination.Start, uint64(0))
 	assert.Equal(t, pagination.End, uint64(100))
 
-	ret, pagination = GetObjectsPage[models.Dag](stor, sender, 50, 100)
+	ret, pagination = storage.GetObjectsPage[models.Dag](st, sender, 50, 100)
 	assert.Equal(t, len(ret), 50)
 	assert.False(t, pagination.HasNext)
 	assert.Equal(t, pagination.Start, uint64(50))
 	assert.Equal(t, pagination.End, uint64(100))
 
-	ret, pagination = GetObjectsPage[models.Dag](stor, sender, 0, 25)
+	ret, pagination = storage.GetObjectsPage[models.Dag](st, sender, 0, 25)
 	assert.Equal(t, len(ret), 25)
 	assert.True(t, pagination.HasNext)
 	assert.Equal(t, pagination.Start, uint64(0))
@@ -99,7 +100,7 @@ func TestGetPaginatedWeekStats(t *testing.T) {
 }
 
 func TestStorage(t *testing.T) {
-	addr := MakeEmptyAddressStats("test")
+	addr := storage.MakeEmptyAddressStats("test")
 	{
 		storage := NewStorage("/tmp/test")
 		if err := storage.addToDBTest(addr, addr.Address, 0); err != nil {
@@ -111,7 +112,7 @@ func TestStorage(t *testing.T) {
 		storage := NewStorage("/tmp/test")
 		defer storage.Close()
 		ret := storage.GetAddressStats("test")
-		if !ret.isEqual(addr) {
+		if !ret.IsEqual(addr) {
 			t.Error("Broken DB")
 		}
 	}
@@ -119,20 +120,20 @@ func TestStorage(t *testing.T) {
 }
 
 func TestCleanStorage(t *testing.T) {
-	stats := MakeEmptyAddressStats("test")
+	stats := storage.MakeEmptyAddressStats("test")
 	stats.PbftCount = 1
-	storage := NewStorage("/tmp/test")
-	defer storage.Close()
+	st := NewStorage("/tmp/test")
+	defer st.Close()
 
-	if err := storage.addToDBTest(stats, stats.Address, 0); err != nil {
+	if err := st.addToDBTest(stats, stats.Address, 0); err != nil {
 		t.Error(err)
 	}
 
-	if err := storage.Clean(); err != nil {
+	if err := st.Clean(); err != nil {
 		t.Error(err)
 	}
 
-	err := storage.getFromDB(stats, getKey(getPrefix(stats), "test", 0))
+	err := st.getFromDB(stats, getKey(getPrefix(stats), "test", 0))
 
 	if err == nil {
 		t.Error("Clean DB does not work")
@@ -141,26 +142,24 @@ func TestCleanStorage(t *testing.T) {
 }
 
 func TestBatch(t *testing.T) {
-	storage := NewStorage("")
-	defer storage.Close()
+	st := NewStorage("")
+	defer st.Close()
 
-	addr := MakeEmptyAddressStats("test")
-	batch := storage.NewBatch()
+	addr := storage.MakeEmptyAddressStats("test")
+	batch := st.NewBatch()
 
 	batch.AddToBatch(addr, addr.Address, 0)
-	addr1 := MakeEmptyAddressStats("test1")
+	addr1 := storage.MakeEmptyAddressStats("test1")
 	batch.AddToBatch(addr1, addr1.Address, 0)
-	if err := batch.Commit(nil); err != nil {
-		t.Error(err)
-	}
+	batch.CommitBatch()
 
-	ret := storage.GetAddressStats("test")
-	if !ret.isEqual(addr) {
+	ret := st.GetAddressStats("test")
+	if !ret.IsEqual(addr) {
 		t.Error("Broken DB")
 	}
 
-	ret = storage.GetAddressStats("test1")
-	if !ret.isEqual(addr1) {
+	ret = st.GetAddressStats("test1")
+	if !ret.IsEqual(addr1) {
 		t.Error("Broken DB")
 	}
 }
