+++ internal/utils/task.go
@@ -1,4 +1,10 @@
-package indexer
+package utils
+
+import (
+	"runtime"
+
+	"github.com/spiretechnology/go-pool"
+)
 
 type Task[P any] struct {
 	f      func(P) error
@@ -43,4 +49,9 @@ func (t *TaskWithResult[P, R]) Run() {
 	if exec_err != nil {
 		*t.err = exec_err
 	}
+}
+
+// isn't creating threads, but limiting goroutines count. Mostly used for RPC and db related tasks
+func MakeThreadPool() pool.Pool {
+	return pool.New(uint(runtime.NumCPU()))
 }
